lint

+++ contracts/interfaces/external/ICryptoPunksMarket.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.4;
 
 /// @title Interface for WETH9
 interface ICryptoPunksMarket {
-    function buyPunk(uint punkIndex) external payable;
+    function buyPunk(uint256 punkIndex) external payable;
 
-    function transferPunk(address to, uint punkIndex) external;
+    function transferPunk(address to, uint256 punkIndex) external;
 }
